customers/requirements.txt
strawberry-graphql==0.137.0
SQLAlchemy==1.4.42
pytest==7.2.0
PyJWT==2.6.0
asyncpg==0.27.0
